requirements.txt
langchain==0.1.16
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube-transcript-api>=0.6.3
unstructured
pytube
numexpr
huggingface_hub
streamlit
yt-dlp
requests